requirements.txt
aenum==3.1.15
aiohttp==3.9.0 ; python_version >= '3.8'
aiosignal==1.3.1 ; python_version >= '3.7'
annotated-types==0.6.0 ; python_version >= '3.8'
attrs==23.1.0 ; python_version >= '3.7'
blinker==1.7.0 ; python_version >= '3.8'
certifi==2023.11.17 ; python_version >= '3.6'
charset-normalizer==3.3.2 ; python_full_version >= '3.7.0'
click==8.1.7 ; python_version >= '3.7'
deprecated==1.2.14 ; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3'
flask==3.0.0
frozenlist==1.4.1 ; python_version >= '3.8'
future==0.18.3 ; python_version >= '2.6' and python_version not in '3.0, 3.1, 3.2, 3.3'
idna==3.6 ; python_version >= '3.5'
itsdangerous==2.1.2 ; python_version >= '3.7'
jinja2==3.1.2 ; python_version >= '3.7'
line-bot-sdk==3.5.1
markupsafe==2.1.3 ; python_version >= '3.7'
multidict==6.0.4 ; python_version >= '3.7'
pydantic==2.5.2 ; python_version >= '3.7'
pydantic-core==2.14.5 ; python_version >= '3.7'
python-dateutil==2.8.2 ; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3'
requests==2.31.0 ; python_version >= '3.7'
six==1.16.0 ; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3'
typing-extensions==4.9.0 ; python_version >= '3.8'
urllib3==2.1.0 ; python_version >= '3.8'
werkzeug==3.0.1 ; python_version >= '3.8'
wrapt==1.16.0 ; python_version >= '3.6'
yarl==1.9.4 ; python_version >= '3.7'
